requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-env==0.2.1
django-filter==21.1
django-phonenumber-field==6.1.0
django-rest-passwordreset==1.2.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
iniconfig==1.1.1
kombu==5.2.4
msgpack==1.0.4
packaging==21.3
phonenumbers==8.12.50
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
